requirements.txt
pytrends
psycopg2
pyyaml
Flask
Flask-cors
flask-restful
flask_apscheduler
python-dotenv